bash: set up sourcing of git prompt function
git provides a shell function that will get the name of the current
branch for use in a shell prompt. On Debian-based systems, it seems to
be sourced by default. This is not the case for Arch, so I had to add
some alternate yadm files to source the file - the file doesn't exist on
my Mint system, so making the alternate files was necessary