Sway: Change how screenshot works, update hwconfig
Changed the main config file so that the screenshot key is set in the
hardware_config file. This way, each host can set its screenshot key to
the key that works for that system e.g. on my laptop, the screenshot key
outputs the Windows shortcut for screenshotting, so I set the
$screenshot-key Sway variable to that shortcut