Notification Scripts: remove function keyword
It seems like the function keyword isn't compliant with the
POSIX shell standard - when I tried the scripts on Linux Mint (to test
portability) they threw an error which got fixed by removing the
function keyword. This wasn't an issue on Arch because the only shells I
had were bash and zsh, and /bin/sh is symlinked to bash, whereas on Linux Mint
I think /bin/sh is symlinked to dash.