Waybar/Sway: Change how waybar is launched
To support having configs for different window managers, I ended up
having to ditch Sway's normal way of launching waybar. This is because
swaybar_command doesn't allow passing of arguments to the command, so I
couldn't pass the argument for a custom config file to waybar.

What I have now is a custom exec_always that re-executes waybar each
time the Sway config is loaded